feat(test-utils): add getElementStyles to test utils
